requirements.txt
aiogram
python-dotenv
loguru
aiohttp
beautifulsoup4 